# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/quotes/send-email": {
      "post": {
        "description": "Endpoint para enviar cotizaciones por email usando Gmail con App Password. Envía un email con formato HTML profesional incluyendo los detalles de la cotización, productos, precios y totales.",
        "operationId": "QuotesController_sendEmail",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendEmailQuoteDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Email enviado correctamente"
          },
          "400": {
            "description": "Solicitud inválida"
          },
          "500": {
            "description": "Error interno del servidor"
          }
        },
        "summary": "Enviar cotización por email",
        "tags": [
          "Quotes"
        ]
      }
    },
    "/api/quotes": {
      "post": {
        "description": "Permite crear una nueva cotización con código único",
        "operationId": "QuotesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuotesCreateRequestDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Cotización creada correctamente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuotesCreateResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Solicitud inválida"
          },
          "404": {
            "description": "Recurso no encontrado"
          },
          "500": {
            "description": "Error interno del servidor"
          }
        },
        "summary": "Crear nueva cotización",
        "tags": [
          "Quotes"
        ]
      },
      "get": {
        "description": "Permite obtener una lista paginada de cotizaciones",
        "operationId": "QuotesController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Cotizaciones encontradas correctamente"
          },
          "400": {
            "description": "Solicitud inválida"
          },
          "500": {
            "description": "Error interno del servidor"
          }
        },
        "summary": "Listar cotizaciones",
        "tags": [
          "Quotes"
        ]
      }
    },
    "/api/quotes/{id}": {
      "get": {
        "description": "Permite obtener una cotización específica con todos sus detalles",
        "operationId": "QuotesController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Cotización encontrada correctamente"
          },
          "400": {
            "description": "Solicitud inválida"
          },
          "404": {
            "description": "Recurso no encontrado"
          },
          "500": {
            "description": "Error interno del servidor"
          }
        },
        "summary": "Obtener cotización por ID",
        "tags": [
          "Quotes"
        ]
      },
      "patch": {
        "description": "Permite actualizar el estado de una cotización (PENDIENTE, ENVIADA, ACEPTADA, RECHAZADA)",
        "operationId": "QuotesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuotesUpdateRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Cotización actualizada correctamente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuotesUpdateResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Solicitud inválida"
          },
          "404": {
            "description": "Recurso no encontrado"
          },
          "500": {
            "description": "Error interno del servidor"
          }
        },
        "summary": "Actualizar cotización",
        "tags": [
          "Quotes"
        ]
      },
      "delete": {
        "description": "Permite marcar una cotización como rechazada",
        "operationId": "QuotesController_delete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Cotización eliminada correctamente"
          },
          "400": {
            "description": "Solicitud inválida"
          },
          "404": {
            "description": "Recurso no encontrado"
          },
          "500": {
            "description": "Error interno del servidor"
          }
        },
        "summary": "Eliminar cotización",
        "tags": [
          "Quotes"
        ]
      }
    },
    "/api/sales": {
      "post": {
        "description": "Permite crear una nueva venta y actualizar automáticamente el stock",
        "operationId": "SalesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SalesCreateRequestDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Venta creada correctamente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SalesCreateResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Solicitud inválida"
          },
          "404": {
            "description": "Recurso no encontrado"
          },
          "500": {
            "description": "Error interno del servidor"
          }
        },
        "summary": "Crear nueva venta",
        "tags": [
          "Sales"
        ]
      },
      "get": {
        "description": "Permite obtener una lista paginada de ventas",
        "operationId": "SalesController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Ventas encontradas correctamente"
          },
          "400": {
            "description": "Solicitud inválida"
          },
          "500": {
            "description": "Error interno del servidor"
          }
        },
        "summary": "Listar ventas",
        "tags": [
          "Sales"
        ]
      }
    },
    "/api/sales/{id}": {
      "get": {
        "description": "Permite obtener una venta específica con todos sus detalles",
        "operationId": "SalesController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Venta encontrada correctamente"
          },
          "400": {
            "description": "Solicitud inválida"
          },
          "404": {
            "description": "Recurso no encontrado"
          },
          "500": {
            "description": "Error interno del servidor"
          }
        },
        "summary": "Obtener venta por ID",
        "tags": [
          "Sales"
        ]
      }
    },
...
  },
  "info": {
    "title": "Sistema de Almacén API",
    "description": "API para gestión de almacén - Productos, Cotizaciones, Ventas y Proveedores",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:4000",
      "description": "Servidor Local"
    }
  ],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Enter JWT token",
        "in": "header"
      }
    },
    "schemas": {
...
      "ItemDto": {
        "type": "object",
        "properties": {
          "nombre": {
            "type": "string",
            "example": "Filtro de aceite",
            "description": "Nombre del producto"
          },
          "cantidad": {
            "type": "number",
            "example": 2,
            "description": "Cantidad de productos"
          },
          "precioUnitario": {
            "type": "number",
            "example": 25.5,
            "description": "Precio unitario del producto"
          },
          "subtotal": {
            "type": "number",
            "example": 51,
            "description": "Subtotal (cantidad × precio unitario)"
          }
        },
        "required": [
          "nombre",
          "cantidad",
          "precioUnitario",
          "subtotal"
        ]
      },
      "SendEmailQuoteDto": {
        "type": "object",
        "properties": {
          "emailDestino": {
            "type": "string",
            "example": "cliente@email.com",
            "description": "Email del cliente al que se enviará la cotización"
          },
          "clienteNombre": {
            "type": "string",
            "example": "Juan Pérez",
            "description": "Nombre completo del cliente"
          },
          "codigo": {
            "type": "string",
            "example": "COT-1234567890-001",
            "description": "Código único de la cotización"
          },
          "fecha": {
            "type": "string",
            "example": "2024-01-15",
            "description": "Fecha de la cotización (formato YYYY-MM-DD)"
          },
          "vendedorNombre": {
            "type": "string",
            "example": "María González",
            "description": "Nombre del vendedor que generó la cotización"
          },
          "total": {
            "type": "number",
            "example": 125.5,
            "description": "Total de la cotización en formato numérico"
          },
          "items": {
            "description": "Lista de productos incluidos en la cotización",
            "example": [
              {
                "nombre": "Filtro de aceite",
                "cantidad": 2,
                "precioUnitario": 25.5,
                "subtotal": 51
              },
              {
                "nombre": "Bujías",
                "cantidad": 1,
                "precioUnitario": 35,
                "subtotal": 35
              }
            ],
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ItemDto"
            }
          }
        },
        "required": [
          "emailDestino",
          "clienteNombre",
          "codigo",
          "fecha",
          "vendedorNombre",
          "total",
          "items"
        ]
      },
      "QuoteItemDto": {
        "type": "object",
        "properties": {
          "productId": {
            "type": "number",
            "example": 1
          },
          "cantidad": {
            "type": "number",
            "example": 2
          },
          "precioUnitario": {
            "type": "number",
            "example": 25.5
          }
        },
        "required": [
          "productId",
          "cantidad",
          "precioUnitario"
        ]
      },
      "QuotesCreateRequestDto": {
        "type": "object",
        "properties": {
          "clienteNombre": {
            "type": "string",
            "example": "Juan Pérez"
          },
          "clienteEmail": {
            "type": "string",
            "example": "juan@email.com"
          },
          "items": {
            "example": [
              {
                "productId": 1,
                "cantidad": 2,
                "precioUnitario": 25.5
              },
              {
                "productId": 2,
                "cantidad": 1,
                "precioUnitario": 15
              }
            ],
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/QuoteItemDto"
            }
          }
        },
        "required": [
          "clienteNombre",
          "clienteEmail",
          "items"
        ]
      },
      "QuotesCreateResponseDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Quote created successfully"
          },
          "id": {
            "type": "number",
            "example": 1
          },
          "codigo": {
            "type": "string",
            "example": "COT-2024-001"
          },
          "clienteNombre": {
            "type": "string",
            "example": "Juan Pérez"
          },
          "clienteEmail": {
            "type": "string",
            "example": "juan@email.com"
          },
          "total": {
            "type": "number",
            "example": 66
          },
          "estado": {
            "type": "string",
            "example": "PENDIENTE",
            "enum": [
              "PENDIENTE",
              "ENVIADA",
              "ACEPTADA",
              "RECHAZADA"
            ]
          },
          "status_code": {
            "type": "number",
            "example": 201
          }
        },
        "required": [
          "message",
          "id",
          "codigo",
          "clienteNombre",
          "clienteEmail",
          "total",
          "estado",
          "status_code"
        ]
      },
      "QuotesUpdateRequestDto": {
        "type": "object",
        "properties": {
          "estado": {
            "type": "string",
            "example": "ENVIADA",
            "enum": [
              "PENDIENTE",
              "ENVIADA",
              "ACEPTADA",
              "RECHAZADA"
            ]
          }
        }
      },
      "QuotesUpdateResponseDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Quote updated successfully"
          },
          "status_code": {
            "type": "number",
            "example": 200
          }
        },
        "required": [
          "message",
          "status_code"
        ]
      },
      "SaleItemDto": {
        "type": "object",
        "properties": {
          "productId": {
            "type": "number",
            "example": 1
          },
          "cantidad": {
            "type": "number",
            "example": 2
          },
          "precioUnitario": {
            "type": "number",
            "example": 25.5
          }
        },
        "required": [
          "productId",
          "cantidad",
          "precioUnitario"
        ]
      },
      "SalesCreateRequestDto": {
        "type": "object",
        "properties": {
          "metodoPago": {
            "type": "string",
            "example": "EFECTIVO",
            "enum": [
              "EFECTIVO",
              "TRANSFERENCIA",
              "TARJETA",
              "CREDITO"
            ]
          },
          "clienteNombre": {
            "type": "string",
            "example": "Juan Pérez"
          },
          "clienteEmail": {
            "type": "string",
            "example": "juan@email.com"
          },
          "clienteTelefono": {
            "type": "string",
            "example": "123456789"
          },
          "cotizacionId": {
            "type": "number",
            "example": 1
          },
          "items": {
            "example": [
              {
                "productId": 1,
                "cantidad": 2,
                "precioUnitario": 25.5
              },
              {
                "productId": 2,
                "cantidad": 1,
                "precioUnitario": 15
              }
            ],
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SaleItemDto"
            }
          }
        },
        "required": [
          "metodoPago",
          "clienteNombre",
          "items"
        ]
      },
      "SalesCreateResponseDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Sale created successfully"
          },
          "id": {
            "type": "number",
            "example": 1
          },
          "fecha": {
            "type": "string",
            "example": "2024-01-15T00:00:00.000Z"
          },
          "total": {
            "type": "number",
            "example": 66
          },
          "metodoPago": {
            "type": "string",
            "example": "EFECTIVO",
            "enum": [
              "EFECTIVO",
              "TRANSFERENCIA",
              "TARJETA",
              "CREDITO"
            ]
          },
          "clienteNombre": {
            "type": "string",
            "example": "Juan Pérez"
          },
          "clienteEmail": {
            "type": "object",
            "example": "juan@email.com"
          },
          "status_code": {
            "type": "number",
            "example": 201
          }
        },
        "required": [
          "message",
          "id",
          "fecha",
          "total",
          "metodoPago",
          "clienteNombre",
          "status_code"
        ]
      }
    }
  }
}